+++ project/src/main/java/com/elevate360/project/controller/dashboard/AdminController.java
@@ -1,8 +1,9 @@
 package com.elevate360.project.controller.dashboard;
-import com.elevate360.project.dto.AssignTraineesRequest;
+
 import com.elevate360.project.entity.User;
 import com.elevate360.project.repo.UserRepository;
 import com.elevate360.project.service.UserService;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -11,7 +12,7 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/admin")
-@CrossOrigin(origins = "http://localhost:4200/")
+@CrossOrigin(origins = "*")
 public class AdminController {
 
     @Autowired
@@ -26,10 +27,12 @@ public class AdminController {
         return userRepository.findByRole("trainer");
     }
 
-    // Get all trainees
+    // Get all trainees with trainer information
     @GetMapping("/trainee")
     public List<User> getAllTrainees() {
-        return userRepository.findByRole("trainee");
+        // Fetch all trainees from the repository
+        List<User> trainees = userRepository.findByRole("trainee");
+        return trainees;
     }
 
     @DeleteMapping("/trainers/{id}")
@@ -69,6 +72,7 @@ public class AdminController {
     }
 
     // Update a trainee by ID
+    @Transactional
     @PutMapping("/trainees/{id}")
     public User updateTrainee(@PathVariable Long id, @RequestBody User updatedTrainee) {
         User trainee = userRepository.findById(id).orElse(null);
@@ -85,10 +89,7 @@ public class AdminController {
         return null;  // If trainee not found or role mismatch
     }
 
-
-//    @Autowired
-//    private UserService userService;
-
+    // Add a new trainer
     @PostMapping("/addtrainer")
     public ResponseEntity<User> addTrainer(@RequestBody User user) {
         user.setRole("trainer");
@@ -96,6 +97,7 @@ public class AdminController {
         return ResponseEntity.ok(savedUser);
     }
 
+    // Add a new trainee
     @PostMapping("/addtrainee")
     public ResponseEntity<User> addTrainee(@RequestBody User user) {
         user.setRole("trainee");
@@ -103,28 +105,36 @@ public class AdminController {
         return ResponseEntity.ok(savedUser);
     }
 
+    // Assign a trainee to a trainer
+    @PutMapping("/trainees/{traineeId}/assign/{trainerId}")
+    public ResponseEntity<User> assignTraineeToTrainer(@PathVariable Long traineeId, @PathVariable Long trainerId) {
+        User trainee = userRepository.findById(traineeId).orElse(null);
+        User trainer = userRepository.findById(trainerId).orElse(null);
 
+        if (trainee != null && trainer != null && "trainee".equals(trainee.getRole()) && "trainer".equals(trainer.getRole())) {
+            // Instead of setting a string, assign the full User (trainer) object
+            trainee.setTrainer(trainer);  // Set the trainer (which is a User object)
 
-//    @PutMapping("/trainees/{traineeId}/assign/{trainerId}")
-//    public ResponseEntity<User> assignTraineeToTrainer(@PathVariable Long traineeId, @PathVariable Long trainerId) {
-//        User trainee = userRepository.findById(traineeId).orElse(null);
-//        User trainer = userRepository.findById(trainerId).orElse(null);
-//
-//        if (trainee != null && trainer != null && "trainee".equals(trainee.getRole()) && "trainer".equals(trainer.getRole())) {
-//            trainee.setTrainer(trainer);
-//            User updatedTrainee = userRepository.save(trainee);
-//            return ResponseEntity.ok(updatedTrainee);
-//        }
-//
-//        return ResponseEntity.status(404).body(null); // Return 404 if trainee or trainer is not found
-//    }
-//    @GetMapping("/trainers/{id}/trainees")
-//    public List<User> getTraineesByTrainer(@PathVariable Long id) {
-//        User trainer = userRepository.findById(id).orElse(null);
-//        if (trainer != null && "trainer".equals(trainer.getRole())) {
-//            return userRepository.findByTrainer(trainer);
-//        }
-//        return null;
-//    }
+            // Save the updated trainee
+            User updatedTrainee = userRepository.save(trainee);
+
+            return ResponseEntity.ok(updatedTrainee);  // Return the updated trainee object
+        }
+
+        return ResponseEntity.status(404).body(null);  // Return 404 if trainee or trainer is not found
+    }
+
+    // Get all trainees assigned to a specific trainer
+    @GetMapping("/trainers/{id}/trainees")
+    public List<User> getTraineesByTrainer(@PathVariable Long id) {
+        // Fetch the trainer by id
+        User trainer = userRepository.findById(id).orElse(null);
+
+        // If the trainer exists and has the role 'trainer'
+        if (trainer != null && "trainer".equals(trainer.getRole())) {
+            // Fetch the list of trainees assigned to this trainer
+            return userRepository.findByTrainer(trainer);  // Return all trainees for this trainer
+        }
+        return null;  // Return null if trainer is not found or role mismatch
+    }
 }
-
